+++ src/signals/incident-management/containers/Dashboard/charts/area-chart.test.ts
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (C) 2023 Gemeente Amsterdam
 import { getAreaChart } from './area-chart'
 
 const mockValues = [
