+++ src/signals/incident-management/containers/Dashboard/components/styled.ts
@@ -1,22 +1,43 @@
 // SPDX-License-Identifier: MPL-2.0
 // Copyright (C) 2023 Gemeente Amsterdam
-import { themeColor } from '@amsterdam/asc-ui'
+import { themeColor, themeSpacing } from '@amsterdam/asc-ui'
 import styled from 'styled-components'
 
-export const Wrapper = styled.span`
+export const Title = styled.h6`
+  font-size: ${themeSpacing(4.5)};
+  line-height: ${themeSpacing(8)};
+  color: ${themeColor('tint', 'level7')};
+  margin-bottom: 0;
+`
+
+export const Subtitle = styled.p`
+  color: ${themeColor('tint', 'level5')};
+  line-height: ${themeSpacing(6)};
+  margin-top: 0;
+  margin-bottom: ${themeSpacing(5)};
+`
+export const Amount = styled.span`
+  color: ${themeColor('primary')};
+`
+
+export const AreaChartWrapper = styled.span`
+  position: relative;
+`
+
+export const ComparisonRateWrapper = styled.span`
   position: absolute;
   display: flex;
   flex-direction: column;
   color: ${themeColor('primary')};
   right: 50px;
   bottom: 60px;
-  font-size: 14px;
-  line-height: 16px;
+  font-size: ${themeSpacing(3.5)};
+  line-height: ${themeSpacing(4)};
   font-weight: 700;
   text-align: right;
 `
 
 export const Description = styled.span`
-  font-size: 7px;
-  line-height: 8px;
+  font-size: ${themeSpacing(1.75)};
+  line-height: ${themeSpacing(2)};
 `
