+++ src/signals/incident-management/containers/Dashboard/styled.ts
@@ -9,24 +9,3 @@ export const StyledRow = styled.div`
   background-color: ${themeColor('tint', 'level2')};
   margin: 0;
 `
-
-export const Title = styled.h6`
-  font-size: 18px;
-  line-height: 32px;
-  color: ${themeColor('tint', 'level7')};
-  margin-bottom: 0;
-`
-
-export const Subtitle = styled.p`
-  color: ${themeColor('tint', 'level5')};
-  line-height: 24px;
-  margin-top: 0;
-  margin-bottom: ${themeSpacing(5)};
-`
-export const Amount = styled.span`
-  color: ${themeColor('primary')};
-`
-
-export const Wrapper = styled.span`
-  position: relative;
-`
